* testsuite/reloc2.sh: If libs aren't built with -fpic, 	disable the test under SELinux in enforcing mode.
git-svn-id: svn://sourceware.org/svn/prelink/trunk@153 94c539fb-cf18-0410-b60f-edeeb537fa16

+++ testsuite/reloc2.sh
@@ -5,6 +5,8 @@ case "`uname -m`" in
   ia64|ppc*|x86_64|mips*) SHFLAGS=-fpic;; # Does not support non-pic shared libs
   s390*) if file reloc1lib1.so | grep -q 64-bit; then SHFLAGS=-fpic; fi;;
 esac
+# Disable this test under SELinux if textrel
+test -z "$SHFLAGS" -a -x /usr/sbin/getenforce -a "`/usr/sbin/getenforce`" = Enforcing && exit 77
 rm -f reloc2 reloc2lib*.so reloc2.log
 $CC -shared $SHFLAGS -O2 -o reloc2lib1.so $srcdir/reloc2lib1.c
 $CC -shared $SHFLAGS -O2 -o reloc2lib2.so $srcdir/reloc2lib2.c \
